RSS/requirements.txt
requests
psycopg2-binary
feedparser
